Remove OpenAiAdapter
